refactor(teams): implement teams retrieval handler with error handling and organization query extraction
