--- server/api/teams.get.ts
@@ -1,18 +1,0 @@
-// Generated by Copilot
-import { defineEventHandler, getQuery } from 'h3';
-import { getTeams } from './teams';
-
-export default defineEventHandler(async (event) => {
-  try {
-    // Generated by Copilot: Extract organization from query parameters
-    const query = getQuery(event);
-    const organization = query.organization as string;
-    
-    // Generated by Copilot: Pass organization to getTeams function
-    const teams = await getTeams(event, organization);
-    return teams;
-  } catch (error) {
-    console.error('Error in teams.get handler:', error);
-    throw error;
-  }
-});
